[CHORE] properties 수정
